requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreschema==0.0.4
Django==2.2.4
django-coreapi==2.3.0
django-cors-headers==3.1.0
django2-rest-passwordreset==1.3.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
